Cartoon writer
# Speedcartoon
Animation part 1
kelvin a young boy,he was very intelligent and experimental trying to produce a machine he could fit in, a robot that can fight to protect his country from war,he knows nothing about his father his mother takes care of him they are not rich, but his mother works as a nurse to provide for him,in their house kelvin has a secret basement where he did his research and experiments, they live in a country called Persia it is a modernized country that have conflicts with other countries, persia is very advance that it uses robots to defend their territory kelvin is 14 years old and is about to finish high school, his mother loved him very much and takes good care of him, kelvin is a very shy boy at school but he is very intelligent he often get bullied at school because he had no strength to fight, so kelvin resolved to create a super intelligent and strong robot  to fight injustice, he hates being bullied at school and loved to help people,he was determine in creating a robot that was strong enough to destroy thousands of army, and he could fit inside the robot, the robot was almost completed but he needed a device that he don't have, he had to steal the device from a lab and he almost got caught kelvin escaped to his house, when Kelvin got home he found his mother dead, she was shot with a bullet, that was the biggest shock he has ever felt in his life kelvin cried, injured himself and was about to commit succide, when he saw a letter written by his mother the letter states, kelvin my precious son i am very sorry for not telling you who your father is,he is a murderer and a wicked criminal i took you away to protect you from him, his enemy are looking for him but can't find him so they came here and shot me i saved money for you in the bank take it and go to Michael your uncle and live there Don't ever come back.kelvin was so angry and he was determined to take revenge for his mother's death he fixed his robot and he was ready for anything. when he finished creating his robot it was  more stronger and intelligent than he expected he named it speed because it was faster than anything in the world and was so fast that it could not be seen when running, speed entered his robot and began learning how to control it during flying and running he had many accident when he was learning how to control his robot but he never gave up, and soon he was ready to become the most strongest hero on the planet able to destroy anything that stand in his way.                     the end of part 1.